+++ scipy/misc/__init__.py
@@ -33,6 +33,7 @@ below are not available without it.
    info - Get help information for a function, class, or module
    lena - Get classic image processing example image Lena
    logsumexp - Compute the log of the sum of exponentials of input elements
+               (imported from scipy.special)
    pade - Pade approximation to function as the ratio of two polynomials
    toimage - Takes a numpy array and returns a PIL image [requires Pillow]
    source - Print function source code
@@ -43,12 +44,12 @@ below are not available without it.
 from __future__ import division, print_function, absolute_import
 
 __all__ = ['who', 'source', 'info', 'doccer',
-           'comb', 'factorial', 'factorial2', 'factorialk']
+           'comb', 'factorial', 'factorial2', 'factorialk', 'logsumexp']
 
 from . import doccer
 from .common import *
 from numpy import who, source, info as _info
-from scipy.special import comb, factorial, factorial2, factorialk
+from scipy.special import comb, factorial, factorial2, factorialk, logsumexp
 
 import sys
