+++ scipy/special/basic.py
@@ -19,6 +19,7 @@ from . import specfun
 from . import orthogonal
 from ._comb import _comb_int
 
+
 __all__ = ['agm', 'ai_zeros', 'assoc_laguerre', 'bei_zeros', 'beip_zeros',
            'ber_zeros', 'bernoulli', 'berp_zeros', 'bessel_diff_formula',
            'bi_zeros', 'clpmn', 'comb', 'digamma', 'diric', 'ellipk',
@@ -2468,3 +2469,4 @@ def zeta(x, q=None, out=None):
     if q is None:
         q = 1
     return _zeta(x, q, out)
+
